Semaphore: refactors - move sending and scheduling stuff around.
